# Adds a pool of UTF-8 decoders, making reader instantiation less expensive.