*Issue #, if available:*
Fixes #370 

*Description of changes:*
1. Created generic base classes for poolable objects (`Poolable` - the negative diff for `Utf8StringEncoder` shows what was extracted) and for pools of poolable objects (`Pool` - the negative diff for `Utf8StringEncoderPool` shows what was extracted).
2. Created `Utf8StringDecoder` (extends `Poolable`) and `Utf8StringDecoderPool` (extends `Pool`) to provide reusable logic and character buffers for decoding UTF-8 strings.
3. Modified `IonReaderBinaryIncremental` (the new incremental binary reader) and `IonReaderBinaryRawX` (the old binary reader) to use `Utf8StringDecoderPool` and `Utf8StringDecoder`.
4. Because `IonReaderBinaryRawX` requires an additional `ByteBuffer` for its string decoding (not required by the incremental reader), created `PoolableByteBuffer` (extends `Poolable`) and `ByteBufferPool` (extends `Pool`) to provide reusable `ByteBuffer`s to this reader.

This takes care of a lot of the expense that comes with instantiating binary readers. To test, I benchmarked a loop like the following on a binary Ion file containing only `"foo"` (using a manual change to [`ion-java-benchmark-cli`](https://github.com/amzn/ion-java-benchmark-cli) that I plan to add as a proper option [at some point](https://github.com/amzn/ion-java-benchmark-cli/issues/12)):

```java
for (int i = 0; i < 10000; i++) {
    IonReader reader = readerBuilder.build(buffer);
    fullyTraverse(reader, false);
    reader.close();
}
```

Results before the proposed change:

```
Benchmark                                               (input)                                                    (options)  Mode  Cnt          Score         Error   Units
Bench.run                                   singleStringFoo.10n      read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:INCREMENTAL}  avgt    3       6136.583 ±    1525.138   us/op
Bench.run:Heap usage                        singleStringFoo.10n      read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:INCREMENTAL}  avgt    3        218.982 ±    4715.263      MB
Bench.run:Serialized size                   singleStringFoo.10n      read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:INCREMENTAL}  avgt    3         ≈ 10⁻⁵                    MB
Bench.run:·gc.alloc.rate                    singleStringFoo.10n      read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:INCREMENTAL}  avgt    3      14709.968 ±    3647.687  MB/sec
Bench.run:·gc.alloc.rate.norm               singleStringFoo.10n      read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:INCREMENTAL}  avgt    3   99440016.278 ±       0.260    B/op
Bench.run:·gc.churn.PS_Eden_Space           singleStringFoo.10n      read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:INCREMENTAL}  avgt    3      14694.721 ±    4046.463  MB/sec
Bench.run:·gc.churn.PS_Eden_Space.norm      singleStringFoo.10n      read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:INCREMENTAL}  avgt    3   99335597.947 ± 3029246.426    B/op
Bench.run:·gc.churn.PS_Survivor_Space       singleStringFoo.10n      read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:INCREMENTAL}  avgt    3          0.144 ±       0.273  MB/sec
Bench.run:·gc.churn.PS_Survivor_Space.norm  singleStringFoo.10n      read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:INCREMENTAL}  avgt    3        971.336 ±    1938.260    B/op
Bench.run:·gc.count                         singleStringFoo.10n      read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:INCREMENTAL}  avgt    3        703.000                counts
Bench.run:·gc.time                          singleStringFoo.10n      read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:INCREMENTAL}  avgt    3        272.000                    ms

Bench.run                                   singleStringFoo.10n  read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:NON_INCREMENTAL}  avgt    3       7526.666 ±     588.629   us/op
Bench.run:Heap usage                        singleStringFoo.10n  read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:NON_INCREMENTAL}  avgt    3        700.571 ±    3701.472      MB
Bench.run:Serialized size                   singleStringFoo.10n  read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:NON_INCREMENTAL}  avgt    3         ≈ 10⁻⁵                    MB
Bench.run:·gc.alloc.rate                    singleStringFoo.10n  read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:NON_INCREMENTAL}  avgt    3      16081.923 ±    1245.981  MB/sec
Bench.run:·gc.alloc.rate.norm               singleStringFoo.10n  read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:NON_INCREMENTAL}  avgt    3  133360016.341 ±       0.232    B/op
Bench.run:·gc.churn.PS_Eden_Space           singleStringFoo.10n  read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:NON_INCREMENTAL}  avgt    3      16022.834 ±    1376.757  MB/sec
Bench.run:·gc.churn.PS_Eden_Space.norm      singleStringFoo.10n  read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:NON_INCREMENTAL}  avgt    3  132869899.349 ± 2851250.571    B/op
Bench.run:·gc.churn.PS_Survivor_Space       singleStringFoo.10n  read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:NON_INCREMENTAL}  avgt    3          0.187 ±       0.541  MB/sec
Bench.run:·gc.churn.PS_Survivor_Space.norm  singleStringFoo.10n  read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:NON_INCREMENTAL}  avgt    3       1552.268 ±    4454.582    B/op
Bench.run:·gc.count                         singleStringFoo.10n  read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:NON_INCREMENTAL}  avgt    3        674.000                counts
Bench.run:·gc.time                          singleStringFoo.10n  read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:NON_INCREMENTAL}  avgt    3        271.000                    ms
```

Results after the change:

```
Benchmark                                               (input)                                                    (options)  Mode  Cnt         Score         Error   Units
Bench.run                                   singleStringFoo.10n      read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:INCREMENTAL}  avgt    3      2492.856 ±     199.448   us/op
Bench.run:Heap usage                        singleStringFoo.10n      read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:INCREMENTAL}  avgt    3       181.488 ±    4739.342      MB
Bench.run:Serialized size                   singleStringFoo.10n      read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:INCREMENTAL}  avgt    3        ≈ 10⁻⁵                    MB
Bench.run:·gc.alloc.rate                    singleStringFoo.10n      read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:INCREMENTAL}  avgt    3      5971.134 ±     476.275  MB/sec
Bench.run:·gc.alloc.rate.norm               singleStringFoo.10n      read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:INCREMENTAL}  avgt    3  16400016.116 ±       0.010    B/op
Bench.run:·gc.churn.PS_Eden_Space           singleStringFoo.10n      read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:INCREMENTAL}  avgt    3      5956.038 ±     873.075  MB/sec
Bench.run:·gc.churn.PS_Eden_Space.norm      singleStringFoo.10n      read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:INCREMENTAL}  avgt    3  16358390.389 ± 1170599.800    B/op
Bench.run:·gc.churn.PS_Survivor_Space       singleStringFoo.10n      read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:INCREMENTAL}  avgt    3         0.195 ±       0.633  MB/sec
Bench.run:·gc.churn.PS_Survivor_Space.norm  singleStringFoo.10n      read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:INCREMENTAL}  avgt    3       536.332 ±    1761.203    B/op
Bench.run:·gc.count                         singleStringFoo.10n      read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:INCREMENTAL}  avgt    3       676.000                counts
Bench.run:·gc.time                          singleStringFoo.10n      read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:INCREMENTAL}  avgt    3       264.000                    ms

Bench.run                                   singleStringFoo.10n  read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:NON_INCREMENTAL}  avgt    3      2611.016 ±     239.516   us/op
Bench.run:Heap usage                        singleStringFoo.10n  read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:NON_INCREMENTAL}  avgt    3        69.573 ±     317.594      MB
Bench.run:Serialized size                   singleStringFoo.10n  read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:NON_INCREMENTAL}  avgt    3        ≈ 10⁻⁵                    MB
Bench.run:·gc.alloc.rate                    singleStringFoo.10n  read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:NON_INCREMENTAL}  avgt    3      3031.101 ±     280.326  MB/sec
Bench.run:·gc.alloc.rate.norm               singleStringFoo.10n  read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:NON_INCREMENTAL}  avgt    3   8720016.116 ±       0.095    B/op
Bench.run:·gc.churn.PS_Eden_Space           singleStringFoo.10n  read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:NON_INCREMENTAL}  avgt    3      3025.731 ±     308.857  MB/sec
Bench.run:·gc.churn.PS_Eden_Space.norm      singleStringFoo.10n  read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:NON_INCREMENTAL}  avgt    3   8704552.518 ±   84175.653    B/op
Bench.run:·gc.churn.PS_Survivor_Space       singleStringFoo.10n  read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:NON_INCREMENTAL}  avgt    3         0.150 ±       0.332  MB/sec
Bench.run:·gc.churn.PS_Survivor_Space.norm  singleStringFoo.10n  read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:NON_INCREMENTAL}  avgt    3       430.269 ±     918.321    B/op
Bench.run:·gc.count                         singleStringFoo.10n  read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:NON_INCREMENTAL}  avgt    3       647.000                counts
Bench.run:·gc.time                          singleStringFoo.10n  read::{f:ION_BINARY,t:BUFFER,a:STREAMING,R:NON_INCREMENTAL}  avgt    3       254.000                    ms
```

I tried with various other file sizes as well, and observed (as expected) that the benefit diminishes as the size of the stream increases and more time is spent on parsing than reader instantiation.

I'm going to continue to experiment with pooling in the readers and writers. At the extreme, we could pool an object that holds _everything_ readers and writers need. I'll determine whether the performance benefits would be worth the complexity.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.