@tgregg 2021-10-08
I reviewed the results of the failing performance regression detector task, and I'm comfortable with them. It detected a very small regression (0.847ms vs 0.860ms, 0.793ms vs 0.803 ms) in two of the three binary read tests, and an improvement in the third (0.926ms vs 0.671ms). In all cases, the slower data point had a super high % error (>40%), indicating either that something may have been competing for resources on the host or that we need to run more iterations, more warmups, throw out runs with a % error above a certain threshold, etc. Note: these tests currently instantiate a single reader, read a small amount of data (~50KB), and close the reader. No performance impact is expected; I confirmed this locally with my ion-java-benchmark-cli runs.